GER-399+GER-396 Persistence of reindex-queue + timezone
The persistence of the reindex queue has been added, and a timezone
comparison is now performed on the reindexing.
This means that as long as the database works on the same timezone of
the underlying system, we don't need to worry about the database
timezone.
So every gerrit can be in their particular timezone, and the database
must be using the same system timezone (i.e. no default_time_zone must
be added in my.cnf)
To accomplish the timezone comparison Gerrit now sends the timezone
offset of the sending node. This is compared with the timezone of the
receiving node.

Change-Id: I0f49be1441a28c41cb266dd89bb76f0174dd1b4a
(cherry picked from commit e066653a5cc7189fc0f1c7cc479a0f2a931c49c8)