GER-420: fix reindex local on deleted index
When an index is deleted locally it should not be reindexed.
There is a queue in ReplicatedIndexEventManager which will host the
latest reindexed changes sent to the other nodes. These changes will be
reindexed after a while to allow for a consistent output of the reindex
job across all the nodes. If a change happens to be deleted, then this
should not cause a NPE while trying to reindex that change which is no
longer in the database.

Change-Id: Iac8a05718b3faa605fa29b4a2c99437815a55869