GER-434 ENV vars that affect curl
We need to check the users system to see if any of the environment
variables are set that will affect curl. If one or more is set,
display a message to the user that the following variables are
set and will affect curl, list set variables and their values,
then ask the user if they want to continue?

Change-Id: I5a5541e066a342147c4ec01b5189f46f11cb8475