gerrit-console-api initial review
The gerrit console api is to be used in conjunction with GitMS
in order to give back a reading of a config file in the All-Projects
directory.

GitMS will use a ProcessBuilder in order to run the jar an get the
required config information. This will be intially used with LFS in order
to read the lfs.config within the All-Projects repo.

Change-Id: Icda5320b7455fbc81ed16b27acd1f1b222a9d4c1