GER-539: Merge of the gerrit console-api.jar project
Gerrit console api now exists to obtain gerrit project information as a standalone jar which can be used from any process.  this keeps additional dependencies out of other projects as it can be called on the command line and returns the configuration file as a whole or sub sections if required.

Change-Id: Ie3c9ca70da3c00d99bd30aa41971c1367ab63689