GER-721: Adding run-integration-tests to Makefile help context
Needed to update the makefile to be informed about whether it was running on jenkins or on a dev box.  so that the clone operations would correctly happen as the git user on the dev box.  Hopefully the git configuration is correct on the jenkins box to return build.jenkins as the user.

Change-Id: I75141338256d96b3efc749a543ecd60deb135313