Refactoring Phase 1 - finding relocated modules.
Finding and updating code, which now points to //java/xxx instead of //gerrit-module-x which it used to.
Update the JGIT_VERS and JGIT_REPO properties to be public, so we can use them in other module builds.
GerritConsoleApi and plugin work has begun, to bring them up to date, adn get them compiling - still not complete.

Change-Id: Id9b7827aeb6d8297a99b374f93eff5d071464552