GER-919: Handling of new events for free, simplification
This change removes wandisco code from existing event types as we now assume
all events are replicated. If events are to be skipped for any reason, there is
a couple of options availble.
    1). @skipReplication annotation is available for programitcally skipping
        and event type
    2). We can now skip events via a configurable in the GitMS application.properties.
        This will be a comma sepearated list of eventNames. We check this list for any events
        that should be skipped from replication.
Changes made to the Replicator class to make use of GitMSApplicationProperties class
as the same code is present in this class and it cuts down on code duplication.

Change-Id: Ia6a424d7f6fc6cd1d5779a88a74a7079491e017d