Asset gathering / setup-environment support scripts.
First support script, setup-environment.sh is used to setup the custom bazel WD repository, and also to select where our temporary assets are grouped.  Whether than be in a jenkins folder on the build machine, or under /tmp/builds/gerritms on a developers box.  Of course it can be overridden using a field GERRIT_TESTING_LOCATION and set to whatever you wish it to be.   Another script list_asset_locations is used to gather the location of the assets into a single property, this enables this to be removed from the jobs we have that uses this.  It also allows us to supply asset paths to other items such as integration tests and archive / deploy steps.

Change-Id: I785f925462beee29d67ed0e786e15624e73bdf33