Allow WORKSPACE info to be supplied as override to list_assets
Allowing the workspace to be suppled means we can control how much of the prefix is chopped off the ASSETS_FOUND strings.   This allows a workspace of /home/jenkins/job1/ to either have the gerrit repo cloned to the workspace root or to a subdirectory as in /home/jenkins/jobs1/gerrit.  So us just finding out where the repository is, is not enough we also need to know where the workspace  root is.

Change-Id: Ice0587ec57ced35ed08e4758aec6348e0e2595be