GER-940: Large rework to resolve ReplicatedEvents processing.
I have no introduced the start of dependency injection of our Replicated classes.  This allows us to greatly reduce the factory type overhead, the singletons are enforced, and we can start to use provided contexts for us like ThreadLocalRequestContext or OneUseRequestContext.
Another simple change was once this was done, I could move from using Provider<ReviewDB> to SchemaFactory<ReviewDb> without changing any of the other classes in the chain as the types are available as injected variables.  This should be done for all our classes as it will make changes much easier in the long run.  It will also remove our footprint in core classes as our objects will be created for free e.g. instead of using addToReplication() we simply bind to the type such that when it is created we are also created, and we can inject the type required.  In this manner we no longer have a footprint.  See example above: ReplicatedEventsManager.hookOnListeners is no longer required or used.

Change-Id: I0c0cc232f0a953a4840d165d79070b5ebdcd8c68