GER-942: Update the ReplicatedIndexEventsManager to use injection.
We need to change up to use the official Providers of SchemaFactory to get hold of a dbProvider fresh every time.  We cannot cache something which has a request context.  Alternative would be to have a sessionContext something like the OneOffRequestContext.  Plus side is we no longer have to require code in the ChangeIndexer to hook ourself in, it is done automatically by the good injection code.

Change-Id: Ie83f17abba7d494621d4412586c0510dbf421ddd