GER-953: Update to Gerrit to support sha1 hashing of project name.
Uses the utility method available in gerrit-gitms-interface to hash the name
of project for the event to a sha1 to avoid collisions with other events that may
share the same timestamp and nodeId if under heavy load.

Changes made to the file filtering, a common class that can be shared between file filters
for replicated events.

Change-Id: I639ba50c947898a9eb0e601f5f89391b62e9e250