GER-964 Fix to console-api standalone jar creation and inclusion
I have added the console-api to be built alongside the main api components for inclusion automatically in their api.zip.  We also include the console-api in our own installer package to I updated the create_installer.sh script to just rename the jar in the scratch directory so we didn't cause any conflict / deletion of the console-api.jar which exists for debugging and for use by console-api the wrapper script. Fix made to preserve the timestamps of the built components being added to the gerritms installer.

Change-Id: I2f5bcdaac594f218651ab5e0c7a2ab0d3148f181