LFS Support as core plugin
Support building LFS from the gerrit workspace, and not only as a standalone project.  As such the LFS jgit references needed to be in our workspace / jgit description so that @jgit-lfs would work....  You can still check this out as its own repository and it will continue to build standalone also!

Change-Id: I374b4a55accb3463636d5174d46e11968b0c0c52