GER-698 Change Next steps output to be printed after Removing obsolete directories during upgrade
As of GER-562, during an upgrade of GerritMS the installer will check
for the existence of some obsolete event directories as well as any
persisted json event files and remove them. At the end of the installation
a list of the removed files will be printed, the issue is that this output
is appearing after the "Next steps" are printed.

Change-Id: I29143a481aa4e91a5e985efa4f11e87fd938af09