GER-1112 : Adding entry to Makefile to run the minimum set of working acceptance tests.
Individual failing JUnit tests have been ignored using @Ignored annotation above them
and recorded on Jira ticket ger-1128.

Tests which are just not relevant such as the replication plugin tests are set to disabled
using the disabled tag in the Build file.

Tests with the tags elastic, docker and disabled are excluded

Change-Id: Iac27d1eb42c296570e065182a252192670d4352f