GER-1197 : Remove older org.codehaus jackson-core from gerrit-gitms-shared
What was changed
We use the older version of jackson-mapper and jackson-core for validation and json conversion.
The new version of Gerrit now removed the use of codehaus and just uses the new lib/jackson com.fasterxml versions.
Updating both Gerrit and gerrit-gitms-interface to use the latest versions of the com.fasterxml versions of jackson.

What was the reason for the fix
The codehaus version of jackson had a security vulnerability.

Change-Id: I5b8a53f2f17141f4aa124a01b72f9ca827a8da5f