GER-1144: Analyse codebase for replication functionality.
What was changed
I've updated the EventTypesTest to include the test testKnownChangeEvents. This test will look at the
registered events and compare it against our known list of events. If any of the events are different or simply
not in the known list then the test assertion will fail.

I've also added a small test to check that the SkipReplication annotation is detected on events.

What was the reason for the fix
The previous way of checking the known events relied on checking the constructor for our
replicated boolean using the reflection API, this is no longer possible as we should be replicating
all events for free now.

Change-Id: I2a2c51c67efce0b6fbf0d83ad097f8da296aa026