GER-1280 Running console-api.jar: help <command> printing empty error message
Improved error reporting to have more content, as to better inform
user what error has occurred.

It was noticed that when testing logging of error messages for the console-api.jar,
that the messages where not sufficiently descriptive enough and lacked any real
explaination of what the problem was, so the work done here was to include the
exception that was thrown in the error message.

Change-Id: I86bf3303be38d9421c160b976de5c039a30c86dc