GER-1299: Pass Java Props to Child Process
Updated code to supply Java properties to child processes

Code now passes along any java params from the parent process
to child processes like the start of the daemon or the reindex
process. This allows init to be run on the release.war locally
wihtout the need for the --no-auto-start or --no-reindex flags
being set on the cmd line.

Change-Id: I403cec82f27c0a6c0a61bbc0fccc1f00419ee6e9