GER-1131: GerritMS template servicefile wrong path
Changing the description in this template to be more
genric as where to store the copy of the template
based on opertaing system.

The rational for this change is due to the fact that
while the "/lib/systemd/system" filepath is valid for
a centos machine, it does not exist in sles machines.

Change-Id: I6e5f401ad2289c346ec81ca3310483780ef1f5ac