GER-955: GerritMS installer reports service not running
Small change on how the error message is printed out.

In the case of a non interactive install of GerritMS,
the installer does not report that the gitms replicator
is not running. In this instance, the user is given no
indication that the install has failed (even less why it
failed). This can easily lead users to think that the
gerritms install was successful when that is not the case.

Change-Id: Ied569203dec3dc5e5e236b4a48853db59d37096e