GER-995: GerritMS non-interactive not showing error's
Added a new function to handle error messages.

In the case of a non interactive install of GerritMS,
the installer does not report any errors that might
have occured during install.  This behavour gives no
indication that the install has failed (even less why
it failed). This can easily lead users to think that
the gerritms install was successful when that is not
the case.  Also on completion of the install the
general information given to the user on how to
complete the installation is also not displayed.

Change-Id: I0f2cfc93e0f45ce1cb5756ea019d749c73519073