GER-1431: Output error message for atomic update
Gerrit rejects remaining updates with a default message
of internal server error, but as we have an UpdateException
to hand we can send the root cause to the console first.

It doesn't make much sense to try and print this per rejection
or even to find a message per exception as the first error
found in GitMS will be the one to fail the whole batch.

+ Get the root cause of the exception caught while handling
  regular commands and output it before dealing with the
  rejections.

Change-Id: I9ac041b8bc5d3aaa462d4a07da1ba48f7faf1cac