GER-1228: Fix confusing success+error on push fail
When GitMS is down, the exceptions thrown are not handled
in a great way, resulting in a mix of success AND failure
messages being printed out to the user.

During testing, a NPE was also spotted during execution of
repro steps, so I have included changes to handle:
* updated messaging to be specific about GitMS being down;
* added try/catch/finally clauses to help fix up various
cases which went on to cause NPE;
* moved error handling up a level - certain methods are
executed for both direct pushes and magic branches - in
the case of direct pushes to patchsets, magicBranch is null
and causes NPE.
* making sure to set status of commands in the event of
failure via reject() to help cleanup the MultiProgressMonitor
 - calls to end() for the subtasks don't register a completion
unless we have set the status.

Change-Id: I971ad461886b3e37f7cb713c1beef64518e2eb08