GER-1493: Project creation caused proposal of index again
ProjectCacheImpl was calling a replicated reindex when
a cache event was received on project creation.

Change is that when generated locally we send the cache event to
everyone and reindex locally with all the others doing a local
reindex on receipt of the event and skipping the replication
of that index call.

Change-Id: I8a3b0a1699bf857c372467a99212ead1b30e8766