GER-1462: Partial revert of changes
Reverting error messaging to previous way.

There exists integration tests which fail to verify the
messages added to receivePack (they may not actually be
written to the stream yet). Reverting to the previous direct
receivePack.sendMessage() for the error cases.
This may be that the tests are bad too but for now, reverting.
The fix still stands for adding normal messaging as this is how
other success messages are printed.

Change-Id: I2fff34f06eac4dad561ed63de46e8b3d83876d32