GER-1462: Delay final message until progress done
Changed use of sendMessage back to addMessage for the cases
when sending the "Update successful" and "GitMS - update
replicated" messages, and also for the error message when
handling an atomic batch update.

When using sendMessage directly there was a chance the
message was being inserted into the the stream in the middle
of progress monitor data, corrupting the monitor.

Change-Id: Ia76857623a2ec32d3893e8be6bbd14a6cb56aaa5