GER-0000: Fix make clean command (re-execution)
Make clean could not be run on a clean system, causing issues when running builds.

If e.g. a make all fails to build an asset, and you try to run make all again, it will
fail as the downloaded-artifacts/*jgit* listing fails.  Alternatively if you have never
run a build which download jgit or gerritgitms assets it would also fail.

Aim is to allow something like this:
  "make clean | make all"

I have updated the make clean target to simply output a message that
no assets are found if this happens.
This allows the make clean target to continue and not exit.

Change-Id: I29b22aa41f415b099b71a090c37b49281bfb8c33