GER-1446: Remove wildcard query from accounts
The wildcard query 'q=*' was added to AccountPredicates
to provide an alternative way to list user accounts but
it appears not to work. Rather than fix it, we already
have a query that does the same job: q=is:active used
by the licensing thread so removing this addition to
AccountPredicates.java as it's one less potential
conflict point when merging vanilla changes.

Change-Id: I0f5e0d914bb7d8084718aecb3ad236bf8a3b993a