GER-1531: Fix for change sequence race condition
In 1.10.0 a workaround for sequencing was added requiring
an offset to be added to sequence batches to avoid two nodes
submitting changes with the same change sequences and GitMS
accepting them on a first-come first-serve basis.

This change makes that no longer necessary by incorporating
node-id:change# in the sequence blob so that two changes with the same
sequence from Gerrit will not be accepted by GitMS as the
blobs will not hash the same.

+ Change the storage/parsing of sequence numbers in RepoSequence.java
  for replicated flow. Store will create a tuple of nodeId:change#
  and parse will check which format it's reading back.
+ Reset the value for initial sequence gap to its vanilla default.
+ Remove the requirement for sequenceBatchSize to be set in the
  config on startup.

Change-Id: I5e3443d14c53a764bcc53f9fa79a7a6fae44f610