GER-0000: Fix skipped event file ordering
Add ability to self heal the event file skipped list.

Basically when this was written we only ever added to the 
skipped event file list as a FIFO. 
But when a failure happens in the worker, the current file 
is always the earliest so should be prepended instead.  

Instead of making more complicated calling code we moved 
order to using a PriorityQueue.
This means every insert into this queue will be correctly ordered 
by the FILE object natural ordering by default 
which is the FILESYSTEM path.
This matches the FILE[] sort which we perform in the 
incoming event worker. 
The ordered queue has to be kept correct so that 
when we skip file 2 and 3 because file 1 was in progress.  
If it fails or completes we can correctly pick up the next
skipped over file.
E.g. File 1 completes, the next file to process is file 2.
If File 1 fails, the next file to process is file 1 after backoff period.

Change-Id: I7c888a4d349d1dc81fb3a9af2b774019804cbf51