GER-1769: DbStale / FailImmediately failure fix
Allow combinations of failures to not impact correct behaviour.

Allow DBStale to keep incrementing counter, and always prevent
move to failed.
FailImmediately flag when set in combination is ignored so we keep
backing off the event file due to the StaleDB entry.
Added small performance fix for isAllMissingEventsSafeToIgnore to exit
early if we can't find an event, or its not safe to ignore.

Change-Id: I731a7207bbc6f15893614d70f3a5c7e31ca90ac4