GER-1769: Addition failure for rewriting event file.
re-write event file with failed events.

Not all cases were successfully calling to check if we needed to remove
correctly processed events from the events file.
Specifially if we are in retryCount > 1 and db was stale and not
immediately - it bumped the failure info, but didn't check if needed
to update the correct entires.  We now do this for all cases.

Updated events file being written messages so that the outgoing has
proposed message, whereas the incoming events being update/written
details that it is removing succeeded events.

Change-Id: Ic5da6c829cf826e8d75774f1c3d9def2a2fdc348